Change assertion to test if race condition happens on thread stopping
When the tracker thread is stopped in Python 3.11 in Github Actions, the last message logged to the INFO channel is not "Monitoring thread ended." but instead "The following components were found: [...]".
This breaks our tests only in Github Actions. My hypothesis is that this happens due to a race, so this commit tests that in CI.

added informative comment